backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas
scikit-learn
pymongo==4.6.3
python-multipart==0.0.9
pydantic==2.6.4
pypdf
pytesseract
Pillow
joblib
nltk
scipy
requests
python-dotenv
